switched to `mz:hierarchy_label` for record importance
Previously, we used a combination of `mz:is_funky` and `mz:tier_locality` to determine if a neighbourhood was important enough to include in the PiP service so that localities that don't traditionally have neighbourhoods don't return them.  The WOF team very recently modified data to include a property named `mz:hierarchy_label` to condense "important" records into 1 property.
