+++ test/test.js
@@ -5,4 +5,4 @@ require ('./components/loadJSONTest.js');
 require ('./components/simplifyGeometryTest.js');
 require ('./components/isActiveRecordTest.js');
 require ('./components/filterOutNamelessRecordsTest.js');
-require ('./components/filterOutUnimportantNeighbourhoodsTest.js');
+require ('./components/filterOutUnimportantRecordsTest.js');
